Resolve #782 -- ClientID corrected (#794)
* Fixed some of the system of PlayerId and ClientId
* Fixed for more than 2 players via HeroSpawn ordering

Refs #782 #778

+++ GameServerLib/Chatbox/Commands/ApCommand.cs
@@ -25,7 +25,7 @@ namespace LeagueSandbox.GameServer.Chatbox.Commands
             }
             else if (float.TryParse(split[1], out var ap))
             {
-                _playerManager.GetPeerInfo(userId).Champion.Stats.AbilityPower.FlatBonus += ap;
+                _playerManager.GetPeerInfo((ulong)userId).Champion.Stats.AbilityPower.FlatBonus += ap;
             }
         }
     }
