+++ GameServerLib/Chatbox/Commands/LevelCommand.cs
@@ -31,7 +31,7 @@ namespace LeagueSandbox.GameServer.Chatbox.Commands
                 }
 
                 var experienceToLevelUp = Game.Map.MapProperties.ExpToLevelUp[lvl - 1];
-                _playerManager.GetPeerInfo(userId).Champion.Stats.Experience = experienceToLevelUp;
+                _playerManager.GetPeerInfo((ulong)userId).Champion.Stats.Experience = experienceToLevelUp;
             }
         }
     }
