+++ GameServerLib/Chatbox/Commands/MobsCommand.cs
@@ -41,7 +41,7 @@ namespace LeagueSandbox.GameServer.Chatbox.Commands
                 .Where(xx => xx.Value.Team == team.ToTeamId())
                 .Where(xx => xx.Value is Minion || xx.Value is Monster);
 
-            var client = _playerManager.GetPeerInfo(userId);
+            var client = _playerManager.GetPeerInfo((ulong)userId);
             foreach (var unit in units)
             {
                  _game.PacketNotifier.NotifyPing(client, new Vector2(unit.Value.X, unit.Value.Y), 0, Pings.PING_DANGER);
