+++ GameServerLib/Packets/PacketHandlers/HandleAttentionPing.cs
@@ -18,7 +18,7 @@ namespace LeagueSandbox.GameServer.Packets.PacketHandlers
 
         public override bool HandlePacket(int userId, AttentionPingRequest req)
         {
-            var client = _playerManager.GetPeerInfo(userId);
+            var client = _playerManager.GetPeerInfo((ulong)userId);
             _game.PacketNotifier.NotifyPing(client, new Vector2(req.X, req.Y), req.TargetNetId, req.Type);
             return true;
         }
